@@ -536,16 +536,16 @@
         "package-local": "npm run pack-adapters-local && vsce package"
     },
     "dependencies": {
-        "glob": "^7.2.0",
-        "slash": "^3.0.0"
+        "glob": "^8.0.3",
+        "slash": "^4.0.0"
     },
     "devDependencies": {
-        "@types/glob": "^7.1.4",
-        "@types/node": "^12.11.7",
+        "@types/glob": "^7.2.0",
+        "@types/node": "^18.0.0",
         "@types/vscode": "^1.47.0",
-        "nerdbank-gitversioning": "^3.4.255",
+        "nerdbank-gitversioning": "^3.5.107",
         "tslint": "^6.1.3",
-        "typescript": "^4.5.5",
-        "vsce": "^2.6.7"
+        "typescript": "^4.7.4",
+        "vsce": "^2.9.2"
     }
 }
